+++ plugins/trace-activation/create_localspan_intercepter.go
@@ -26,12 +26,20 @@ type CreateLocalSpanInterceptor struct {
 }
 
 func (h *CreateLocalSpanInterceptor) BeforeInvoke(invocation operator.Invocation) error {
-	operationName := invocation.Args()[0].(string)
-	s, err := tracing.CreateLocalSpan(operationName)
-	invocation.DefineReturnValues(s, err)
 	return nil
 }
 
 func (h *CreateLocalSpanInterceptor) AfterInvoke(invocation operator.Invocation, result ...interface{}) error {
+	operationName := invocation.Args()[0].(string)
+	s, err := tracing.CreateLocalSpan(operationName)
+	if err != nil {
+		invocation.DefineReturnValues(nil, err)
+		return nil
+	}
+	enhancced, ok := result[0].(operator.EnhancedInstance)
+	if !ok {
+		return nil
+	}
+	enhancced.SetSkyWalkingDynamicField(s)
 	return nil
 }