+++ plugins/trace-activation/instrument.go
@@ -49,88 +49,82 @@ func (i *Instrument) VersionChecker(version string) bool {
 func (i *Instrument) Points() []*instrument.Point {
 	return []*instrument.Point{
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("CreateEntrySpan"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStructEnhance("SpanRef"),
+		},
+		{
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("CreateEntrySpan"),
 			Interceptor: "CreateEntrySpanInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("CreateLocalSpan"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("CreateLocalSpan"),
 			Interceptor: "CreateLocalSpanInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("CreateExitSpan"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("CreateExitSpan"),
 			Interceptor: "CreateExitSpanInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("StopSpan"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("StopSpan"),
 			Interceptor: "StopSpanInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("CaptureContext"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("CaptureContext"),
 			Interceptor: "CaptureContextInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("ContinueContext"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("ContinueContext"),
 			Interceptor: "ContinueContextInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("GetTraceID"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("GetTraceID"),
 			Interceptor: "GetTraceIDInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("GetSegmentID"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("GetSegmentID"),
 			Interceptor: "GetSegmentIDInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("GetSpanID"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("GetSpanID"),
 			Interceptor: "GetSpanIDInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("AddLog"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewMethodEnhance("*SpanRef", "SetTag"),
+			Interceptor: "AsyncTagInterceptor",
+		},
+		{
+			PackagePath: "", PackageName: "trace", At: instrument.NewMethodEnhance("*SpanRef", "AddLog"),
+			Interceptor: "AsyncLogInterceptor",
+		},
+		{
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("AddLog"),
 			Interceptor: "AddLogInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("SetTag"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("SetTag"),
 			Interceptor: "SetTagInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("SetOperationName"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("SetOperationName"),
 			Interceptor: "SetOperationNameInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("PrepareAsync"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewMethodEnhance("*SpanRef", "PrepareAsync"),
 			Interceptor: "PrepareAsyncInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("AsyncFinish"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewMethodEnhance("*SpanRef", "AsyncFinish"),
 			Interceptor: "AsyncFinishInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("GetCorrelation"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("GetCorrelation"),
 			Interceptor: "GetCorrelationInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("SetCorrelation"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("SetCorrelation"),
 			Interceptor: "SetCorrelationInterceptor",
 		},
 		{
-			PackagePath: "", PackageName: "trace",
-			At:          instrument.NewStaticMethodEnhance("SetComponent"),
+			PackagePath: "", PackageName: "trace", At: instrument.NewStaticMethodEnhance("SetComponent"),
 			Interceptor: "SetComponentInterceptor",
 		},
 	}