+++ CHANGES.md
@@ -12,6 +12,7 @@ Release Notes.
 #### Documentation
 
 #### Bug Fixes
+* Fix users can not use async api in toolkit-trace 
 
 #### Issues and PR
 - All issues are [here](https://github.com/apache/skywalking/milestone/197?closed=1)